refactor(bridge): 패키지를 나눠서 구조 변경
