+++ src/main/java/bridge/Application.java
@@ -1,5 +1,9 @@
 package bridge;
 
+import bridge.controller.InputController;
+import bridge.domain.BridgeGame;
+import bridge.view.InputView;
+
 public class Application {
 
     public static void main(String[] args) {
