+++ src/main/java/bridge/domain/BridgeGame.java
@@ -1,5 +1,9 @@
-package bridge;
+package bridge.domain;
 
+import bridge.BridgeMaker;
+import bridge.BridgeNumberGenerator;
+import bridge.BridgeRandomNumberGenerator;
+import bridge.controller.InputController;
 import java.util.List;
 
 /**
