+++ src/main/java/bridge/domain/Player.java
@@ -1,5 +1,7 @@
-package bridge;
+package bridge.domain;
 
+import bridge.view.InputView;
+import bridge.view.OutputView;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -62,7 +64,6 @@ public class Player {
         if (!isSuccess()) {
             OutputView.printResult("실패", moveCount);
         }
-
     }
 
     private void validateMoving(String moving) {
